data_goverance_langchain/requirements.txt
langchain>=0.1.14
langchain-community>=0.0.19
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
